add rotate value to CutFrameType [#155]

+++ domain/src/main/java/com/foke/together/domain/interactor/entity/PhotoPosition.kt
@@ -5,5 +5,5 @@ data class PhotoPosition (
     val height: Int,
     val x: Int,
     val y: Int,
-    val rotation: Int = 0
+    val rotation: Float = 0f
 )
